split off makeMassDict functions from create_update_header_config
